+++ admin/templates/form.php
@@ -61,7 +61,7 @@
             </table>
         </div><!-- .awsm-form-section-main -->
         <div class="awsm-form-footer">
-            <?php submit_button(); ?>
+        <?php echo apply_filters( 'awsm_job_settings_submit_btn', get_submit_button(), 'form' ); ?>
         </div><!-- .awsm-form-footer -->
     </form>
 </div><!-- .awsm-admin-settings -->