+++ admin/templates/specifications.php
@@ -113,7 +113,7 @@
             </div><!-- .awsm-form-section -->
         </div><!-- .awsm-form-section-main -->
         <div class="awsm-form-footer">
-            <?php submit_button(); ?>
+            <?php echo apply_filters( 'awsm_job_settings_submit_btn', get_submit_button(), 'specifications' ); ?>
         </div><!-- .awsm-form-footer -->
     </form>
 </div><!-- .awsm-admin-settings -->