+++ inc/class-awsm-job-openings-filters.php
@@ -3,10 +3,11 @@ if( ! defined( 'ABSPATH' ) ) {
     exit;
 }
 
-class AWSM_Job_Openings_Filters extends AWSM_Job_Openings {
+class AWSM_Job_Openings_Filters {
     private static $_instance = null;
 
     public function __construct( ) {
+        $this->cpath = untrailingslashit( plugin_dir_path( __FILE__ ) );
         add_action( 'awsm_filter_form', array( $this, 'display_filter_form' ) );
         add_action( 'wp_ajax_jobfilter', array( $this,'awsm_posts_filters' ) );
         add_action( 'wp_ajax_nopriv_jobfilter',  array( $this,'awsm_posts_filters') );
@@ -60,7 +61,7 @@ class AWSM_Job_Openings_Filters extends AWSM_Job_Openings {
             }
         }
 
-        $args = $this->awsm_job_query_args( $filters );
+        $args = AWSM_Job_Openings::awsm_job_query_args( $filters );
 
         if( isset( $_POST['paged'] ) ) {
             $args['paged'] = intval( $_POST['paged'] ) + 1;
@@ -69,7 +70,7 @@ class AWSM_Job_Openings_Filters extends AWSM_Job_Openings {
         $query = new WP_Query( $args );
 
         if ( $query->have_posts() ) :
-            include_once plugin_dir_path( __FILE__ ) . 'templates/partials/listing-view.php';
+            include_once $this->cpath . '/templates/partials/listing-view.php';
         else :
             if( $_POST['action'] !== 'loadmore' ) :
         ?>