+++ inc/templates/job-openings-view.php
@@ -2,16 +2,16 @@
 	if( ! defined( 'ABSPATH' ) ) {
 		exit;
 	}
-	$args = $this->awsm_job_query_args();
+	$args = AWSM_Job_Openings::awsm_job_query_args();
     $query = new WP_Query( $args );
 	if( $query->have_posts() ) : ?>
 		<div class="awsm-job-wrap">
 
 			<?php do_action( 'awsm_filter_form' );  ?>
 
-			<div class="<?php echo $this->get_job_listing_view_class(); ?>" id="awsm-job-response">
+			<div class="<?php echo AWSM_Job_Openings::get_job_listing_view_class(); ?>" id="awsm-job-response">
 				<?php
-					include_once plugin_dir_path(__FILE__) .  'partials/listing-view.php';
+					include_once untrailingslashit( plugin_dir_path( __FILE__ ) ) .  '/partials/listing-view.php';
 				?>
 			</div>