+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## Changelog
 
+### V 1.0.1 - 2018-08-22
+* Added job specifications display options on job detail page appearance settings
+* Added job status meta box on application detail screen
+* Minor fixes and improvements
+
 ### V 1.0 - 2018-08-12
 * Initial Release